Context 0.2: Exporting decoders that may be shared between Entropy and Context
