+++ crates/binjs_io/src/bytes/mod.rs
@@ -6,6 +6,9 @@ pub mod bool;
 /// Compressing/decompressing from/to common formats.
 pub mod compress;
 
+/// Decoders for streams of floats, etc.
+pub mod decoders;
+
 /// Encoding/decoding floating-point numbers.
 pub mod float;
 
@@ -14,6 +17,9 @@ pub mod lengthwriter;
 
 /// Serializing/deserializing traits.
 pub mod serialize;
+
+/// String utilities.
+pub mod strings;
 
 /// Encoding/decoding variable-length numbers.
 pub mod varnum;
