+++ crates/binjs_io/src/entropy/util.rs
@@ -1,6 +1,6 @@
 //! Miscellaneous utilities used for entropy (de)coding.
 
-use std::io::{Cursor, Read};
+use std::io::Read;
 
 /// Utility: an extension of `Read` with a method `expect` that may be
 /// used to ensure the presence of a sequence of bytes.
@@ -28,24 +28,5 @@ where
             }
         }
         Ok(())
-    }
-}
-
-/// Utility: An extension of `Read` that reads `bytes` bytes and attempts
-/// to conver them into a string.
-pub trait ReadStr {
-    fn read_string(&mut self, bytes: usize) -> Result<String, std::io::Error>;
-}
-impl ReadStr for Cursor<Vec<u8>> {
-    fn read_string(&mut self, bytes: usize) -> Result<String, std::io::Error> {
-        use std::io::*;
-        let mut buf = Vec::with_capacity(bytes);
-        unsafe {
-            buf.set_len(bytes);
-        }
-        self.read_exact(&mut buf)?;
-        let result = String::from_utf8(buf)
-            .map_err(|_| Error::new(ErrorKind::InvalidData, "Invalid UTF-8 data"))?;
-        Ok(result)
     }
 }
